+++ src/components/MapView.vue
@@ -13,8 +13,8 @@
       :boundary-padding="10"
     >
       <section class="popover-children">
-        <Items v-if="selection.type === 'items'" :stac="stac" :items="selection.children" />
-        <Catalogs v-else-if="selection.type === 'collections'" collectionsOnly enforceCards :stac="stac" :catalogs="selection.children" />
+        <Items v-if="selection.type === 'items'" :stac="stac" :items="selection.children" showControls enforceView="cards" />
+        <Catalogs v-else-if="selection.type === 'collections'" collectionsOnly enforceView="cards" :stac="stac" :catalogs="selection.children" />
         <Features v-else :features="selection.children" />
       </section>
       <div class="text-center">